requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

Django>=4.2,<5.0
gunicorn>=21.2,<24.0

numpy>=1.24,<2.0
Pillow>=10.0,<12.0
scikit-image>=0.21,<0.24
scipy>=1.10,<1.14
tqdm>=4.66,<5.0
einops>=0.7,<0.9
opencv-python-headless>=4.9,<5.0

torch==2.2.2
torchvision==0.17.2
